Ask for plot factory in `MultiFarming`
